// ClientApp/app/Models/MandateRegInWordHistoricalMandate/HistoricalMandateOnUsClass.ts
export class HistoricalMandateOnUsClass {
    dataList:any;
    ParentBankName: string;
    CorpName: string;
    ActivityName: string;
    MSGId: string;

    MandateFreshId: string;
    mandateMode: string;
    AutoRejectReason: string;
    AcceptRefNo: string;
    updatedon: string;
    UpdateBy: string;
    username: string;
    Enach: string;
    IsMobileData: string;
    RejectedReason: string;
    REJECTED: string;
    IsLiveInNACH: string;
    CreatedOn: string;
    Srno: string;
    IsScan: string;
    IsPrint: string;
    DateOnMandate: string;
    ToDebit: string;
    BankName: string;
    Code: string;
    Frequency: string;
    status: string;
    Customer1: string;
    debittype: string;
    AcNo: string;
    mandateid: string;
    Refrence1: string;
    FromDate: string;
    Amount: string;
    AmountRupees: string;
    EmailId: string;
    SponsorbankCode: string;
    UtilityCode: string;
    PhoneNumber: string;
    NPCIErrorDesc: string;
    ToDate: string;

}